requirements.txt
attrs==21.4.0
beautifulsoup4==4.9.3
blinker==1.5
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.5
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.3.0.post0
pycparser==2.21
pymongo==4.6.0
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
Werkzeug==2.2.3
wsproto==1.2.0